allowed stack to grow up or down
Added a toggle in the StackView component to toggle the growsUp state value. This value controls the 'up' or 'down' class on the stack-content div, which switches the flex-direction from column to column-reverse.  Similar toggles are applied on several elements that descend from the .up or .down ancestor.
To make this work with decoded values that must align with the proper byte, I also refactored the decode() function in decode.js into decode() and format().  The former works as before, decoding a fixedInt value to one of hex, int, uint, char, or bin, but without inserting any line breaks.  The format function handles the line breaks, and takes an additional boolean 'flip' parameter to determine if the string should be reversed before the line breaks are applied.
Also moved a few old stack-related css selectors from style.css into stack.css

Some groundwork for supporting customizable stack chunking, but it is still a work in progress.