updated runtime to force an aligned stack on exec()
Force the exec() function in runtime.js to allocate a multiple of 8 bytes for the argv strings.  This prevents the weirdness of an unaligned stack after executing run(), and makes sure that the stack pointer indicator is visible immediately in the debugger.  TODO: handle unaligned stack better anyway.
Also removed some console.log()s in stdlib