Fixed double label bug
Previously, the assembler maintained a mapping from addresses (proxy: line numbers) to labels, so if two consecutive labels referred to the same address, the second one would overwrite the first.  Now, the assembler maps from line numbers to a list of labels, and the same change was made in vm/Process.js.  Additionally, the InstructionView component in app/TextView.jsx now renders all labels in the list for a particular address.