Added component to display the heap
ProcessView now contains an extra boolean in its state, determining whether or not the heap should be shown.  The heap's visibility is toggled with a tab-like button that appears at the bottom of the window when hidden.  Toggling the heap will squash the two rows of the heapless display to take up the space of one row, and the heap gets the entire second row.

The heap component itself is very much modeled after the stack, as a scrollable window taking up the entire width of the screen, composed of decodable 'heap-items'.  There is quite a bit of similar code between the two that could likely be abstracted out at some point, will add an issue for this.  It's also fairly inefficient on a first render, and it locks up the screen for an appreciable half second or so.  There is a neat 'window' feature, where the current area of the heap being viewed is displayed as a window over a miniature representation of the entire heap, which can be moved to change the focused area.