made memory resizable, added support for heap, brk, sbrk
Finally got around to fixing the resize() function in MemorySegment.  Resizability is now determined by a function canResize passed as the fourth parameter to the MemorySegment constructor.  The stack has canResize() set to a new function resizeDown() which allows it to grow as needed.  MemorySegment.resize() can also be called explicitly by the Memory wrapper function, and this is how Memory.brk and Memory.sbrk are implemented.